Refactor: Optimize shortcut config handling during data sync
This commit refactors how shortcut configurations (`EblanShortcutConfig`) are processed during data synchronization to improve performance and reduce redundant database operations.

The `updateEblanShortcutConfigs` logic, previously called individually for each application during the sync, has been moved out of the per-application loop. Instead, a complete list of all `EblanShortcutConfig`s is now generated first. This list is then compared against the existing data in the database in a single, bulk operation at the end of the sync process.

### Key Changes:

*   **Bulk Processing:** Instead of fetching and updating shortcut configs one app at a time, the `SyncDataUseCase` now accumulates all `EblanShortcutConfig`s from all apps into a single list (`newEblanShortcutConfigs`).
*   **Single Database Transaction:** The comparison, upsert, and deletion of shortcut configs now happen only once per sync, drastically reducing the number of database reads and writes.
*   **Code Consolidation:**
    *   The `updateEblanShortcutConfigs` function was moved from the now-deleted `LauncherAppsUtil.kt` and is now a private function within `SyncDataUseCase` and `ChangePackageUseCase`.
    *   In `SyncDataUseCase`, the original logic was split into two new helper functions: `getEblanShortcutConfigs` to build the list and a new `updateEblanShortcutConfigs` to perform the final database transaction.
*   **Notification Handling:** The progress notification logic in `SyncDataUseCase` was slightly simplified, removing an unnecessary helper function (`notifyProgress`).
